[dev] delete todo by id service gives error for non existing id

+++ src/main/java/com/tdd/todo/service/impl/TodoServiceImpl.java
@@ -59,6 +59,9 @@ public class TodoServiceImpl implements TodoService {
 
     @Override
     public TodoResponse deleteById(UUID id) {
+        Optional<Todo> optionalTodo = todoRepository.findById(id);
+        if (optionalTodo.isEmpty())
+            throw new EntityNotFoundException("Todo not found");
         todoRepository.deleteById(id);
         return null;
     }
